requirements.txt
pandas
psycopg2-binary
jupyter
pydrive2
dvc[gdrive]
dvc[all]
pyarrow
